test(cron): add sinon timers for test cron
This commit add sinon's fake timers to test the cron jobs.
