+++ package.json
@@ -145,6 +145,7 @@
     "protractor": "^5.4.2",
     "qs": "^6.9.1",
     "release-it": "^12.3.5",
+    "sinon": "^7.5.0",
     "snyk": "^1.210.0",
     "standard-version": "^7.0.0",
     "typescript": "^3.3.1"
