+++ test/server-unit/cron/addDynamicDatesOptions.spec.js
@@ -1,0 +1,94 @@
+const { expect } = require('chai');
+const rewire = require('@ima-worldhealth/rewire');
+
+describe('cronEmailReport', () => {
+  let addDynamicDatesOptions;
+
+  const DAILY = 1;
+  const WEEKLY = 2;
+  const MONTHLY = 3;
+  const YEARLY = 4;
+
+  describe('#addDynamicDatesOptions', () => {
+    before(() => {
+      const cronEmailReport = rewire('../../../server/controllers/admin/cronEmailReport');
+      addDynamicDatesOptions = cronEmailReport.__get__('addDynamicDatesOptions');
+    });
+
+
+    it('#addDynamicDatesOptions() skips if hasDynamicDates is false', () => {
+      const options = { id : 1, label : 'Some ole options' };
+      const processed = addDynamicDatesOptions(1, false, options);
+      expect(processed).to.deep.equal(options);
+    });
+
+    it('#addDynamicDatesOptions() does nothing if cronId is unrecognized', () => {
+      const options = { id : 1, label : 'Some ole options' };
+      const processed = addDynamicDatesOptions(123, true, options);
+      expect(processed).to.deep.equal(options);
+    });
+
+    it('#addDynamicDatesOptions() sets the DAILY schedule to the current day', () => {
+      const options = { id : 1, label : 'A schedule' };
+      const processed = addDynamicDatesOptions(DAILY, true, options);
+      expect(processed.dateFrom).to.be.a('object');
+      expect(processed.dateTo).to.be.a('object');
+
+      const today = new Date().toDateString();
+
+      expect(processed.dateFrom.toDate().toDateString()).to.equal(today);
+      expect(processed.dateTo.toDate().toDateString()).to.equal(today);
+    });
+
+    it('#addDynamicDatesOptions() sets the WEEKLY schedule to the current week', () => {
+      const options = { id : 1, label : 'A schedule' };
+
+      const { dateFrom, dateTo } = addDynamicDatesOptions(WEEKLY, true, options);
+      expect(dateFrom).to.be.a('object');
+      expect(dateTo).to.be.a('object');
+
+      const today = new Date();
+
+      expect(dateFrom.toDate().getDay()).to.equal(0);
+      expect(dateTo.toDate().getDay()).to.equal(6);
+
+      expect(dateFrom.toDate().getMonth()).to.equal(today.getMonth());
+      expect(dateTo.toDate().getMonth()).to.equal(today.getMonth());
+    });
+
+
+    it('#addDynamicDatesOptions() sets the MONTHLY schedule to the current month', () => {
+      const options = { id : 1, label : 'A schedule' };
+
+      const { dateFrom, dateTo } = addDynamicDatesOptions(MONTHLY, true, options);
+      expect(dateFrom).to.be.a('object');
+      expect(dateTo).to.be.a('object');
+
+      const today = new Date();
+      expect(dateFrom.toDate().getMonth()).to.equal(today.getMonth());
+      expect(dateTo.toDate().getMonth()).to.equal(today.getMonth());
+
+      expect(dateFrom.toDate().getDate()).to.equal(1);
+      expect(dateTo.toDate().getDate()).to.be.at.least(28);
+    });
+
+    it('#addDynamicDatesOptions() sets the YEARLY schedule to the current year', () => {
+      const options = { id : 1, label : 'A schedule' };
+
+      const { dateFrom, dateTo } = addDynamicDatesOptions(YEARLY, true, options);
+      expect(dateFrom).to.be.a('object');
+      expect(dateTo).to.be.a('object');
+
+      const today = new Date();
+      expect(dateFrom.toDate().getFullYear()).to.equal(today.getFullYear());
+      expect(dateTo.toDate().getFullYear()).to.equal(today.getFullYear());
+
+      expect(dateFrom.toDate().getMonth()).to.equal(0);
+      expect(dateTo.toDate().getMonth()).to.equal(11);
+
+      expect(dateFrom.toDate().getDate()).to.equal(1);
+      expect(dateTo.toDate().getDate()).to.equal(31);
+
+    });
+  });
+});
